requirements.txt
absl-py==2.4.0
astunparse==1.6.3
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.16.0
idna==3.11
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.9
ml-dtypes==0.3.2
MouseInfo==0.1.3
namex==0.1.0
numpy==1.26.4
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
protobuf==3.20.3
PyAutoGUI==0.9.54
pycparser==3.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
requests==2.32.5
rich==14.3.3
rubicon-objc==0.5.3
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sounddevice==0.5.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.2
